cleaning around
removing old code from this branch and making the new code be under
regular filenames without 'ast-' prefix
