--- src/ast-parser/index.js
@@ -1,6 +1,0 @@
-// @flow
-import Statement from './statement'
-
-export default function parse(blob: string) {
-	return new Statement(blob)
-}
