--- src/evaluator/utils.js
@@ -1,29 +1,0 @@
-// @flow
-import { max } from '../utils'
-import type { Tree, TreeItem } from './../parser'
-import { OperatorNode } from './../parser/node'
-
-export const maxOperator = (tree: Tree) => {
-	const res = max(
-		tree,
-		(prev, node) => (
-			!(prev instanceof OperatorNode)
-			|| (node instanceof OperatorNode
-				&& node.getOrder() > prev.getOrder())
-		),
-	)
-
-	return res instanceof OperatorNode && res
-}
-
-export const isSameTreeItem = (a: TreeItem, b: TreeItem) => a === b
-
-// const isSameTrees = (a: Tree, b: Tree) => (
-// 	a.every((item, i) => b[i] && isSameTreeItem(item, b[i]))
-// )
-//
-// export const isSameTreeItem = (a: TreeItem, b: TreeItem) => (
-// 	a instanceof Node
-// 	? a.equals(b)
-// 	: Array.isArray(b) && isSameTrees(a, b)
-// )
