+++ src/parser/statement.js
@@ -1,141 +1,123 @@
 // @flow
-import type { TreeItemType } from './types'
-import * as tt from './types'
-import UtilParser, { getMatch } from './util'
-import Expression from './expression'
-import isNotValidFunction from './../functions'
-import { Node, ArgumentNode, OperatorNode } from './node'
+import * as N from '../types'
+import * as tt from '../tokenizer/types'
+import toTokens from '../tokenizer'
+import type { Token } from '../tokenizer'
+import { getMatch } from '../utils'
+import Node from './node'
+import pushItemToNode from './util'
 
-type Options = {
-	blob?: string,
-	isNewTree?: bool,
-	start?: number,
-	end?: number
+type StatementState = {
+	pos: number,
+	errors: Array<{
+		error: string,
+		pos: number
+	}>
 }
 
-export default class Statement extends UtilParser {
-	tree: Expression
+export default class StatementParser {
 	blob: string
-
-	constructor(blob: string, params?: Set<string>) {
-		super()
-
-		this.tree = new Expression({ params })
-		this.blob = blob
-
-		this.parse()
+	tokens: Token[]
+	tree: N.Expression
+	state: StatementState = {
+		pos: 0,
+		errors: [],
 	}
 
-	parse({
-		blob = this.blob,
-		isNewTree = false,
-		start = 0,
-		end = blob.length - 1,
-	}: Options = {}): Expression {
-		const tree = isNewTree ? new Expression({ parent: this.tree }) : this.tree
-		let pos = start
+	constructor(blob: string) {
+		this.blob = blob
 
-		while (pos <= end) {
-			pos += this.nextNode(pos, blob, tree, pos === start)
+		this.tree = new Node('EXPRESSION', blob, this.state.pos)
+
+		this.parse(toTokens(blob))
+	}
+
+	parse(tokens: Token[], tree: N.Expression = this.tree) {
+		const start = this.state.pos
+		for (const token of tokens) {
+			this.state.pos += this.nextToken(token, tokens, tree)
 		}
+
+		this.state.pos = start
 
 		return tree
 	}
 
-	nextNode(pos: number, blob: string, tree: Expression, firstItem: bool): number {
-		if (blob[pos] === ' ') return getMatch(blob, /\s+/).length
+	nextToken(token: Token, tokens: Token[], tree: N.Expression) {
+		if (this.state.pos >= this.blob.length) return 0
 
-		const { type, match } = this.inferTypeAndMatch(pos, blob)
+		const node = this.parseToken(token)
 
-		// if there now operator between two items, than a '*' operator will be added
-		if (
-			type !== tt.OPERATOR
-			// could be bigger than 0, but it doesn't matter so this is faster
-			&& tree.items.length > 1
-			&& !(tree.get(-1) instanceof OperatorNode)
-		) {
-			tree.add(new OperatorNode(tt.OPERATOR, '*'))
-		}
+		// eslint-disable-next-line no-param-reassign
+		tree.body = tree.body != null ? pushItemToNode(node, tree.body) : node
 
-		tree.add(this.parseToItem(pos, type, match, tree.params, firstItem))
-
-		// if there are two operators near each other
-		if (type === tt.OPERATOR && tree.get(-2) instanceof OperatorNode) {
-			this.unexpected('two operators can not be near each other')
-		} else if (type === tt.PARAM && !tree.params.has(match)) {
-			tree.params.add(match)
-		}
-
-		return type === tt.ERROR || tree.get(-1).type === tt.ERROR ? 1 : match.length
+		return token.match.length
 	}
 
-	parseToItem(
-		pos: number,
-		type: TreeItemType,
-		match: string,
-		params: Set<string>,
-		firstItem: bool,
-	) {
-		switch (type) {
-			case tt.OPERATOR:
-				return this.parseOperatorNode(type, match, pos, firstItem)
-			case tt.NUMBER:
-			case tt.CONSTANT:
-			case tt.PARAM:
-			case tt.ERROR:
-				return new Node(type, match)
-			case tt.FUNCTION:
-				return this.parseFunction(type, match)
-			case tt.ABS_BRACKETS:
-				return this.parseAbsBrackets(type, match, pos)
+	parseToken(token: Token): N.Node {
+		switch (token.type) {
+			case tt.LITERAL:
+				return this.parseLiteral(token)
+			case tt.BIN_OPERATOR:
+				return this.parseBinOperator(token)
 			case tt.BRACKETS:
-				return this.parseBrackets(type, match, pos)
+				return this.parseBrackets(token.match.slice(1, -1))
+			case tt.ABS_BRACKETS:
+				return this.parseAbsBrackets(token)
+			case tt.FUNCTION:
+				return this.parseFunction(token.match)
+			case tt.CONSTANT:
+				return this.parseConstant(token)
 			default:
-				throw new Error(`${type} is not a valid type`)
+				this.state.errors.push({
+					error: `${token.type}: wrong type`,
+					pos: this.state.pos,
+				})
+				return new Node(tt.ERROR, this.state.pos)
 		}
 	}
 
-	parseOperatorNode(type: TreeItemType, match: string, pos: number, firstItem: bool) {
-		const op: OperatorNode = new OperatorNode(type, match)
-
-		// + and - can be in pos 0
-		if (firstItem && op.getOrder() !== 1) {
-			this.unexpected(`${match} can't be the first operator`)
-		}
-
-		return op
+	parseLiteral(token: Token): N.Literal {
+		const node: N.Literal = new Node(token.type, token.match, this.state.pos)
+		node.value = Number(node.raw)
+		return node
 	}
 
-	parseFunction(type: TreeItemType, match: string): Node {
-		const name = getMatch(match, /^[a-z]+/)
-		const args = getMatch(match, /\(.+\)/)
-			.slice(1, -1)
-			.split(',')
-			.map(blob => this.parse({ blob, isNewTree: true }).toArray())
-
-		const isNotValid = isNotValidFunction(name, args)
-
-		if (isNotValid) {
-			this.unexpected(isNotValid)
-		}
-
-		return new ArgumentNode(type, name, args)
+	parseBinOperator(token: Token): N.BinOperator {
+		const node: N.BinOperator = new Node(token.type, token.match, this.state.pos)
+		node.operator = node.raw
+		return node
 	}
 
-	parseAbsBrackets(type: TreeItemType, match: string, pos: number): ArgumentNode {
-		return new ArgumentNode(
-			tt.FUNCTION,
-			'abs',
-			[this.parse({ start: pos + 1, end: pos + match.length - 2, isNewTree: true }).toArray()],
+	parseBrackets(match: string): N.Expression {
+		return this.parse(
+			toTokens(match),
+			new Node('EXPRESSION', match, this.state.pos),
 		)
 	}
 
-	parseBrackets(type: TreeItemType, match: string, pos: number): Expression {
-		// slicing for removing the breckets from the string
-		return this.parse({
-			isNewTree: true,
-			start: pos + 1,
-			end: pos + match.length - 2,
-		})
+	parseAbsBrackets(token: Token): N.Function {
+		const node: N.Function = new Node(tt.FUNCTION, token.match, this.state.pos)
+		node.name = 'abs'
+		node.arguments = [this.parseBrackets(token.match)]
+
+		return node
+	}
+
+	parseFunction(match: string): N.Function {
+		const node: N.Function = new Node(tt.FUNCTION, match, this.state.pos)
+		node.name = getMatch(match, /[a-z]+/)
+		node.arguments = getMatch(match, /\(.+\)/)
+			.slice(1, -1)
+			.split(',')
+			.map(str => this.parseBrackets(str))
+
+		return node
+	}
+
+	parseConstant(token: Token): N.Constant {
+		const node: N.Constant = new Node(token.type, token.match, this.state.pos)
+		node.name = token.match
+		return node
 	}
 }
